+++ app/src/main/java/com/giraffe/myweatherapp/presentation/model/DayUiModel.kt
@@ -5,6 +5,6 @@ import com.giraffe.myweatherapp.R
 data class DayUiModel(
     val dayName: String = "Monday",
     val iconRes: Int = R.drawable.day_mainly_clear,
-    val highTemperature: Int = 32,
-    val lowTemperature: Int = 20,
+    val highTemperature: Double = 32.0,
+    val lowTemperature: Double = 20.0,
 )
