+++ app/src/main/java/com/giraffe/myweatherapp/presentation/model/HourUiModel.kt
@@ -4,6 +4,6 @@ import com.giraffe.myweatherapp.R
 
 data class HourUiModel(
     val iconRes: Int = R.drawable.day_mainly_clear,
-    val temperature: Int = 25,
+    val temperature: Double = 25.0,
     val time: String = "11:00",
 )
